requirements.lock.windows.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.lock.windows.txt requirements.in
#
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==4.4.2
    # via
    #   -r requirements.in
    #   moviepy
fonttools==4.61.1
    # via matplotlib
idna==3.11
    # via requests
imageio==2.37.2
    # via moviepy
imageio-ffmpeg==0.6.0
    # via
    #   -r requirements.in
    #   moviepy
joblib==1.5.3
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.8
    # via -r requirements.in
moviepy==1.0.3
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   imageio
    #   matplotlib
    #   moviepy
    #   opencv-python
    #   pandas
    #   scikit-learn
    #   scipy
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via matplotlib
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via
    #   imageio
    #   matplotlib
proglog==0.1.12
    # via moviepy
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   -r requirements.in
    #   moviepy
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   moviepy
    #   proglog
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
